+++ test/cluster/strong_consistency_test.go
@@ -448,81 +448,6 @@ var _ = Describe("SCMode", func() {
 
 			Expect(DeployCluster(k8sClient, ctx, aeroCluster)).To(HaveOccurred())
 		})
-
-		It("Should not allow ForceBlockFromRoster enable in more than 1 rack at once", func() {
-			By("Deploy")
-
-			aeroCluster := createDummyAerospikeCluster(clusterNamespacedName, 4)
-			aeroCluster.Spec.AerospikeConfig = getSCAerospikeConfig()
-			scNamespace := scNamespace
-			racks := []asdbv1.Rack{
-				{ID: 1},
-				{ID: 2},
-				{ID: 3},
-			}
-			rackConf := asdbv1.RackConfig{
-				Namespaces: []string{scNamespace},
-				Racks:      racks,
-			}
-			aeroCluster.Spec.RackConfig = rackConf
-
-			Expect(DeployCluster(k8sClient, ctx, aeroCluster)).ToNot(HaveOccurred())
-
-			aeroCluster.Spec.RackConfig.Racks[0].ForceBlockFromRoster = ptr.To(true)
-			aeroCluster.Spec.RackConfig.Racks[1].ForceBlockFromRoster = ptr.To(true)
-
-			err := updateCluster(k8sClient, ctx, aeroCluster)
-			Expect(err).Should(HaveOccurred())
-			Expect(err.Error()).To(ContainSubstring(
-				"the forceBlockFromRoster flag can be applied to only one rack at a time"))
-		})
-
-		It("Should not allow rack aware features with ForceBlockFromRoster", func() {
-			By("Deploy")
-
-			aeroCluster := createDummyAerospikeCluster(clusterNamespacedName, 4)
-			aeroCluster.Spec.AerospikeConfig = getSCAerospikeConfig()
-			scNamespace := scNamespace
-			racks := []asdbv1.Rack{
-				{ID: 1},
-				{ID: 2},
-			}
-			rackConf := asdbv1.RackConfig{
-				Namespaces: []string{scNamespace},
-				Racks:      racks,
-			}
-			aeroCluster.Spec.RackConfig = rackConf
-
-			Expect(DeployCluster(k8sClient, ctx, aeroCluster)).ToNot(HaveOccurred())
-
-			aeroCluster.Spec.RackConfig.Racks[0].ForceBlockFromRoster = ptr.To(true)
-			aeroCluster.Spec.RackConfig.RollingUpdateBatchSize = &intstr.IntOrString{IntVal: 2}
-
-			err := updateCluster(k8sClient, ctx, aeroCluster)
-			Expect(err).Should(HaveOccurred())
-			Expect(err.Error()).To(ContainSubstring(
-				"with only one rack in roster, cannot use rollingUpdateBatchSize or scaleDownBatchSize"))
-
-			aeroCluster.Spec.RackConfig.RollingUpdateBatchSize = nil
-			aeroCluster.Spec.RackConfig.MaxIgnorablePods = &intstr.IntOrString{IntVal: 2}
-			err = updateCluster(k8sClient, ctx, aeroCluster)
-			Expect(err).Should(HaveOccurred())
-			Expect(err.Error()).To(ContainSubstring(
-				"forceBlockFromRoster cannot be used together with maxIgnorablePods"))
-
-			aeroCluster.Spec.RackConfig.MaxIgnorablePods = nil
-			aeroCluster.Spec.RosterNodeBlockList = []string{"1A0", "1A1"}
-			err = updateCluster(k8sClient, ctx, aeroCluster)
-			Expect(err).Should(HaveOccurred())
-			Expect(err.Error()).To(ContainSubstring("forceBlockFromRoster cannot be used together with RosterNodeBlockList"))
-
-			aeroCluster.Spec.RosterNodeBlockList = nil
-			aeroCluster.Spec.RackConfig.Racks[1].ForceBlockFromRoster = ptr.To(true)
-			err = updateCluster(k8sClient, ctx, aeroCluster)
-			Expect(err).Should(HaveOccurred())
-			Expect(err.Error()).To(ContainSubstring(
-				"all racks cannot have forceBlockFromRoster enabled. At least one rack must remain in the roster"))
-		})
 	})
 })
